+++ capi/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
   `vmlinux`
 - Added support for disabling `kallsyms` and `vmlinux` to
   `blaze_symbolize_src_kernel`
+- Added `blaze_symbolize_cache_process` for caching process VMA metadata
 
 
 0.1.0-rc.2